Resurrect support for render / refresher loadable libraries and use them to implement  an experimental OpenGL 3.2 renderer. Please note that the new renderer interface is somewhat different from the original one, old render libraries will NOT work!

+++ src/client/cl_cin.c
@@ -348,7 +348,7 @@ Huff1Decompress(cblock_t in)
 
 	if ((input - in.data != in.count) && (input - in.data != in.count + 1))
 	{
-		Com_Printf("Decompression overread by %i", (input - in.data) - in.count);
+		Com_Printf("Decompression overread by %i", (int)(input - in.data) - in.count);
 	}
 
 	out.count = out_p - out.data;
